requirements.txt
annotated-types==0.5.0
attrs==23.1.0
click==8.1.5
delta-spark==2.4.0
et-xmlfile==1.1.0
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
MarkupSafe==2.1.3
numpy==1.25.1
openpyxl==3.1.2
pandas==2.0.3
py4j==0.10.9.7
pydantic==2.0.3
pydantic_core==2.3.0
pyspark==3.4.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
rpds-py==0.8.10
six==1.16.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
yetl-framework==2.0.5.dev2
zipp==3.16.2
